Designed and updated activity_category.xml with categories (not all categories are included due to display not loading all the textViews properly)
