--- app/src/main/java/edu/utsa/cs3443/project/data.java
@@ -1,4 +1,0 @@
-package edu.utsa.cs3443.project;
-
-public class data {
-}
