+++ app/src/main/res/values/strings.xml
@@ -29,4 +29,8 @@
     <string name="Food_input">Enter amount: $</string>
     <string name="Savings_text">Savings:</string>
     <string name="Savings_input">Enter amount: $</string>
+    <string name="income_question_TextView">do_you_want_to_enter_your_monthly_or_annual_salary_m_for_monthly_and_y_for_yearly</string>
+    <string name="Income_question_TextView">Do you want to enter your monthly or yearly salary?</string>
+    <string name="income_TextView">Do you want to enter your monthly (M) or yearly (Y) salary?</string>
+    <string name="enter_amount">Enter amount:</string>
 </resources>
